Update queens.cpp
Optimized fen() function by adding an array to the board which stores active queens. When a scan succeeds, we add it's respective queen into the array, once finished we place the initial queen at the correct location in this array. During the fen processing we can look at our list of queens specifically and create a fen according to the y value. x value denotes place in the array.